# app/Http/Controllers/Api/apiController.php
<?php

namespace App\Http\Controllers\Api;

use App\Http\Controllers\Controller;
use Illuminate\Http\Request;
use App\Models\User;
use Illuminate\Support\Facades\Auth;
use Illuminate\Support\Facades\Hash;
use Illuminate\Support\Facades\Validator;

   /**
     * @OA\Info(
     *     title="Laravel API",
     *     version="1.0.0",
     *     description="API documentation for the Laravel project"
     * )
     *
     * @OA\Server(
     *     url="http://127.0.0.1:8000/api",
     *     description="API v1"
     * )
     */

class apiController extends Controller
{

    /**
     * @OA\Post(
     *     path="/user/register",
     *     summary="Register a new user",
     *     tags={"User"},
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             @OA\Property(property="name", type="string", example="John Doe"),
     *             @OA\Property(property="email", type="string", example="john.doe@example.com"),
     *             @OA\Property(property="password", type="string", example="password")
     *         )
     *     ),
     *     @OA\Response(response=200, description="Success"),
     *     @OA\Response(response=400, description="Bad Request")
     * )
     */

    // Register function
    public function register(Request $request)
    {
        try{
           // Add Validation Multiap type
           $validation = Validator::make($request->all(), [
            'name' => 'required',
            'email' => 'required|email|unique:users,email',
            'password' => 'required',
        ]);

        // Validation All Data
        if ($validation->fails()) {
            return response()->json([
                'status' => false,
                'message' => 'Validation error',
                'errors' => $validation->errors()->all(),
            ],401);
        }
        
        //Response Data in Json
        $users = User::create([
            'name'=>$request->name,
            'email'=>$request->email,
            'password' => Hash::make($request->password),
            'role_id'=>$request->role_id,
            
        ]);

        //Response Data in Json
        return response()->json([
            'status' => true,
            'message' => 'User created successfully',
        ],200);

        }catch(\Exception $e)
        {
            return response()->json([
                'status' => false,
                'message' => 'An error occurred Register',
                'error' => $e->getMessage(),
            ]);
        }
    }

    // User Login
 /**
     * @OA\post(
     *     path="/user/login",
     *     summary="User Login",
     *     tags={"User"},
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             @OA\Property(property="email", type="string", example="john.doe@example.com"),
     *             @OA\Property(property="password", type="string", example="password")
     *         )
     *     ),
     *     @OA\Response(response=200, description="Success"),
     *     @OA\Response(response=400, description="Bad Request")
     * )
     */
   public function login(Request $request)
    {
        try{

            $validation = Validator::make($request->all(), [
                'email'=> 'required|email',
                'password'=> 'required',
            ]);
            if($validation->fails())
            {
                return response()->json([
                    'status' => false,
                    'message' => 'Validation error',
                    'errors' => $validation->errors()->all(),
                ],401);
            }
            if(!Auth::attempt($request->only('email','password')))
            {
                return response()->json([
                    'status' => false,
                    'message' => 'Emial & password Dose not match With Out record.',
                ],401);
            }
            $user = User::Where('email',$request->email)->first();
            return response()->json([
                'status' => true,
                'message' => 'User login is successfully',
                'token' => $user->createToken('AuthToken')->plainTextToken,
            ], 200);

        }catch(\Exception $e)
        {
            return response()->json([
                'ststus'=>false,
                'message' => 'An error occurred in Login',
                'error' => $e->getMessage(),
            ]);
        }
    }

    // User Profile

/**
 * @OA\Get(
 *     path="/user/profile",
 *     tags={"User"},
 *     summary="Fetch authenticated user details",
 *     security={{"bearerAuth":{}}},
 *     @OA\Response(
 *         response=200,
 *         description="User details",
 *         @OA\JsonContent(
 *             @OA\Property(property="id", type="integer"),
 *             @OA\Property(property="name", type="string"),
 *             @OA\Property(property="email", type="string"),
 *         )
 *     ),
 *     @OA\Response(
 *         response=401,
 *         description="Unauthenticated"
 *     )
 * )
 */


    public function profile()
    {
       try{
        $user = auth()->user();
        return response()->json([
         'status'=>true,
         'message'=>'Successfull Fetch data in Profile',
         'data'=>$user
        ]);
 
       }catch(\Exception $e)
       {
           return response()->json([
               'status' => false,
               'message' => 'An error occurred Post',
               'error' => $e->getMessage(),
           ]);
       }
    }

     // User Logout
     public function logout(Request $request)
     {
        try{
            // Get the authenticated user
            $user = $request->user();
                        
            // Revoke all tokens for the user
            $user->tokens()->delete();

            return response()->json(['message' => 'Logged out successfully from all devices'], 200);
        }catch(\Exception $e)
        {
            return response()->json([
                'status' => false,
                'message' => 'An error occurred Post',
                'error' => $e->getMessage(),
            ]);
        }
     }

     // User Role 
     
}